v6/agentic-service/requirements.txt
# Agentic Service Requirements
# Python 3.11+

# Google Agent Development Kit (ADK)
google-adk>=1.18.0

# FastAPI & Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Streamlit for Demo UI
streamlit>=1.29.0
streamlit-option-menu>=0.3.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
numpy>=1.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
backoff>=2.2.0

# Logging
python-json-logger>=2.0.0

# MCP (Model Context Protocol) Support
mcp>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
